Feature/fix-ui

See merge request hcm25_cpl_js_java_02/group1/laboratory-management!88

+++ backend/src/main/java/com/example/test_order_service/serviceImpl/CommentServiceImpl.java
@@ -84,6 +84,11 @@ public class CommentServiceImpl implements CommentService {
         Comment comment = commentRepository.findByCommentIdAndTestOrder_TestOrderId(commentId, orderId)
                 .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
 
+        String currentUsername = GeneralUtils.getCurrentUsername();
+        if (!comment.getCreatedBy().equals(currentUsername)) {
+            throw new IllegalStateException("User does not have permission to update this comment");
+        }
+
         // Luôn cập nhật text mới từ request
         comment.setCommentText(request.getCommentText());
         comment.setUpdatedBy(GeneralUtils.getCurrentUsername());
@@ -120,6 +125,11 @@ public class CommentServiceImpl implements CommentService {
 
         Comment comment = commentRepository.findByCommentIdAndTestOrder_TestOrderId(commentId, orderId)
                 .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
+
+        String currentUsername = GeneralUtils.getCurrentUsername();
+        if (!comment.getCreatedBy().equals(currentUsername)) {
+            throw new IllegalStateException("User does not have permission to delete this comment");
+        }
 
         // Xóa khỏi DB
         commentRepository.deleteById(comment.getCommentId());
